hihi

+++ models/UserModel.php
@@ -2,6 +2,37 @@
 
 class UserModel extends Model {
 
+    public static function findUserInfos($id) {
+        $stmt = Model::connect()->prepare("SELECT * FROM users WHERE id=?");
+        $stmt->execute([$id]);
+
+        $datas = $stmt->fetch(PDO::FETCH_ASSOC);
+
+        if ($datas == null) {
+            header("Location: index.php?error=usernotfound");
+            exit();
+        } else {
+
+            // Recherche sql dans la DB pour savoir si l'user a une photo de profil
+            $stmt = Model::connect()->query("SELECT * FROM users_gallery WHERE id_user='$id'");
+            // On fait le while de recherche
+            while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
+                // Si on voit que l'user à un id similaire dans la table users_gallery, on insère son image dans $_SESSION
+                if ($row['status'] == 1) {
+                    $fileName = "public/img/profile/profile".$id."*";
+                    $fileInfo = glob($fileName);
+                    $fileExt = explode('.', $fileInfo[0]);
+                    $fileActualExt = $fileExt[1];
+                    $datas['picture'] = "public/img/profile/profile".$id.".".$fileActualExt;
+                // Sinon on upload une img par défaut
+                } else {
+                    $datas['picture'] = "public/img/profile/profiledefault.jpg";
+                }
+                return $datas;
+            }
+        }
+    }
+
     public static function signup($username, $email, $password, $passwordRepeat) {
 
         // Si on a rencontré aucun problème, on regarde si un user avec ce nom ou un email n'existent pas déjà
@@ -60,24 +91,24 @@ class UserModel extends Model {
             $stmt = Model::connect()->prepare("SELECT * FROM users WHERE username=? OR email=?;");
             $stmt->execute([$mailuid, $mailuid]);
 
-            $datas = $stmt->fetchAll(PDO::FETCH_ASSOC);
+            $datas = $stmt->fetch(PDO::FETCH_ASSOC);
 
             if ($datas == null) {
                 header("Location: index.php?error=usernotfound");
                 exit();
             // Si on trouve des résultats, on check le password
             } else {
-                $pwdCheck = password_verify($password, $datas['0']['password']);
+                $pwdCheck = password_verify($password, $datas['password']);
                 // Si l'user s'est trompé de mdp, on renvoie une erreur
                 if ($pwdCheck == false) {
                     header("Location: index.php?error=invalidpassword&mailuid=".$mailuid);
                     exit();
                 // Si le mdp est correct, on lance la session et on donne comme id/username les params trouvés dans la db
                 } else {
-                    $_SESSION['userId'] = $datas['0']['id'];
-                    $_SESSION['userUid'] = $datas['0']['username'];
-                    $_SESSION['userEmail'] = $datas['0']['email'];
-                    $_SESSION['role'] = $datas['0']['role'];
+                    $_SESSION['userId'] = $datas['id'];
+                    $_SESSION['userUid'] = $datas['username'];
+                    $_SESSION['userEmail'] = $datas['email'];
+                    $_SESSION['role'] = $datas['role'];
 
                     $id = $_SESSION['userId'];
                     // Recherche sql dans la DB pour savoir si l'user a une photo de profil
@@ -135,14 +166,14 @@ class UserModel extends Model {
 
             // On check si l'extension est autorisée
             if (!in_array($fileActualExt, $allowed)) {
-                header("Location: index.php?action=profile&error=badtype");
+                header("Location: index.php?action=profile&id=".$id."&error=badtype");
                 exit();
             } else {
                 // On check s'il y a eu une erreur
                 if ($fileError === 0) {
                     // On check que la taille ne soit pas trop grande
                     if ($fileSize > 1000000) {
-                        header("Location: index.php?action=profile&error=toobig");
+                        header("Location: index.php?action=profile&id=".$id."&error=toobig");
                         exit();
                     } else {
                         // On peut créer un id aléatoire "XXXXXX.XXXX.extensionfichier"
@@ -160,41 +191,47 @@ class UserModel extends Model {
                         // On modifie l'image de profile sur l'utisateur log
                         $stmt = Model::connect()->query("UPDATE users_gallery SET status=1 WHERE id_user='$id';");
                         $_SESSION['picture'] = $fileDest;
-                        header("Location: index.php?action=profile&upload=success");
+                        header("Location: index.php?action=profile&id=".$id."&upload=success");
                         exit();
                     }
                 } else {
-                    header("Location: index.php?action=profile&error=unknown");
+                    header("Location: index.php?action=profile&id=".$id."&error=unknown");
                     exit();
                 }
             }
+        } else {
+            header("Location: index.php");
         }
     }
 
     public static function deletePicture() {
 
-        $id = $_SESSION['userId'];
+        if(isset($_POST['uploadimg_submit'])) {
+            $id = $_SESSION['userId'];
 
-        // On indique le nom du fichier
-        $fileName = "public/img/profile/profile".$id.".*";
-        // Si on cherche des fichiers avec un id de 1, alors l'image des id 10/11/etc seront aussi prises
-        $fileInfo = glob($fileName);
-        // Pour contrer ça on met [0] après $fileInfo pour n'avoir qu'un résultat (le bon)
-        $fileExt = explode('.', $fileInfo[0]);
-        // L'extension est ici stockée dans le 3, on la récupère
-        $fileActualExt = $fileExt[1];
-        // On a normalement le nom du fichier entier
-        $file = "public/img/profile/profile".$id.".".$fileActualExt;
+            // On indique le nom du fichier
+            $fileName = "public/img/profile/profile".$id.".*";
+            // Si on cherche des fichiers avec un id de 1, alors l'image des id 10/11/etc seront aussi prises
+            $fileInfo = glob($fileName);
+            // Pour contrer ça on met [0] après $fileInfo pour n'avoir qu'un résultat (le bon)
+            $fileExt = explode('.', $fileInfo[0]);
+            // L'extension est ici stockée dans le 3, on la récupère
+            $fileActualExt = $fileExt[1];
+            // On a normalement le nom du fichier entier
+            $file = "public/img/profile/profile".$id.".".$fileActualExt;
 
-        // On supprime le fichier existant
-        if (!unlink($file)) {
-            header("Location: index.php?action=profile&error");
-            exit();
+            // On supprime le fichier existant
+            if (!unlink($file)) {
+                header("Location: index.php?action=profile&id=".$id."&error");
+                exit();
+            }
+            // On indique que l'utilisateur n'a plus de photo de profil perso
+            $stmt = Model::connect()->query("UPDATE users_gallery SET status=0 WHERE id_user='$id';");
+            $_SESSION['picture'] = "public/img/profile/profiledefault.jpg";
+
+            header("Location: index.php?action=profile&id=".$id."&delete=success");
+        } else {
+            header("Location: index.php");
         }
-        // On indique que l'utilisateur n'a plus de photo de profil perso
-        $stmt = Model::connect()->query("UPDATE users_gallery SET status=0 WHERE id_user='$id';");
-        $_SESSION['picture'] = "public/img/profile/profiledefault.jpg";
-
-        header("Location: index.php?action=profile&delete=success");
     }
 }
